fix: [PN] Import d'un PN, en fonction de la version et suppression des anciennes datas

+++ src/Controller/superAdministration/ImportDiplomeController.php
@@ -14,6 +14,7 @@ use App\Classes\Structure\DiplomeImport;
 use App\Controller\BaseController;
 use App\Entity\Constantes;
 use App\Entity\Diplome;
+use App\Repository\PpnRepository;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -25,13 +26,20 @@ class ImportDiplomeController extends BaseController
      * @throws \Exception
      */
     #[Route(path: '/{id}', name: 'sa_diplome_import_index')]
-    public function index(MyUpload $myUpload, DiplomeImport $diplomeImport, Request $request, Diplome $diplome): Response
+    public function index(
+        PpnRepository $ppnRepository,
+        MyUpload $myUpload, DiplomeImport $diplomeImport, Request $request, Diplome $diplome): Response
     {
         if ($request->isMethod('POST')) {
-            $fichier = $myUpload->upload($request->files->get('fichier'), 'temp/', ['xml']);
-            $diplomeImport->import($diplome, $fichier, $request->request->get('typeFichier'));
-            unlink($fichier);
-            $this->addFlashBag(Constantes::FLASHBAG_SUCCESS, 'Maquette importée avec succès');
+            $ppn = $ppnRepository->find($request->request->get('ppn'));
+            if ($ppn !== null) {
+                $fichier = $myUpload->upload($request->files->get('fichier'), 'temp/', ['xml']);
+                $diplomeImport->import($diplome, $fichier, $request->request->get('typeFichier'), $ppn);
+                unlink($fichier);
+                $this->addFlashBag(Constantes::FLASHBAG_SUCCESS, 'Maquette importée avec succès');
+            } else {
+                $this->addFlashBag(Constantes::FLASHBAG_ERROR, 'PPN non trouvé');
+            }
         }
 
         return $this->render('super-administration/import_diplome/index.html.twig', [
